+++ src/ParserErrorType.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rowbot\URL;
+
+enum ParserErrorType
+{
+    case NONE;
+
+    case URL;
+
+    case BASE;
+}